// src/model/address.model.ts
export class AddressResponse{
        id: number;
        street?: string;
        city?: string;
        province?: string;
        country: string;
        postal_code: string;
}

export class CreateAddressRequest{
    contact_id: number;
    street?: string;
    city?: string;
    province?: string;
    country: string;
    postal_code: string;
}